requirements.txt
colorama==0.4.6
keyboard==0.13.5
MouseInfo==0.1.3
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
tqdm==4.67.1
